// lib/api/generated/types/Post.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Status } from "./Status.ts";
import type { User } from "./User.ts";

export type Post = {
    /**
     * @type number
    */
    id: number;
    /**
     * @type string
    */
    title: string;
    /**
     * @type string
    */
    slug: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @type string
    */
    content: string;
    /**
     * @type string
    */
    featuredImage: string;
    /**
     * @type boolean
    */
    isPublished: boolean;
    /**
     * @type string, date-time
    */
    publishedAt: string;
    /**
     * @type object
    */
    author: User;
    /**
     * @type object
    */
    status: Status;
    /**
     * @type string, date-time
    */
    createdAt: string;
    /**
     * @type string, date-time
    */
    updatedAt: string;
    /**
     * @type string, date-time
    */
    deletedAt: string;
};